Basic Map
Right now I have some javascript arrays for the positions and
descriptions, but I think for the future we should have a geojson file
with all this info. I am looking into how we can do that, but google
maps can easily read geojson data so that shouldn’t be a problem. The
bigger problem is that we would want a list of the the clubs in the
body of the page where people can click on it and the marker associated
with that club opens up. This can probably be done with a dom event
listener but I am not as familiar with that.
http://stackoverflow.com/questions/18333679/google-maps-open-info-window
-after-click-on-a-link this page probably has the answers but I don’t
have time to do that right now so yeah