progress
So data.js is just how I’m storing position info right now, definitely
will change in the future. index is for displaying the info. This is
what will eventually be displayed on the school site. table_placer will
let us click on a spot so people can put tables wherever they want. All
I need to know now is how we should write that info to a database.
Also, we need a form so that people can write the description. So maybe
we have a page setup where people click on the spot they want, then
fill out a description form, click submit, and it writes to the
database.